pyproject.toml
[tool.poetry]
name = "clickup-cli"
version = "0.1.0"
description = "CLI tool to test ClickUp API"
authors = ["Nicola Brisotto"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
typer = "^0.12.3"
pydantic-settings = "^2.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
